Load the currencies from the backend and make the displayed values dynamic
We need to connect the FE to our own BE that calls the cnb website.
This is because of a CORS issue with CNB.  CNB only allows CORS if the
domain is in apl.cnb.cz, therefore since we're not on that domain, our
browser blocks it.

We made our own backend, which (for now) allows CORS with any domain
(including ours!).

This change:
- Fetches the currencies from the backend
- Shows the conversion rate as ... before the loading is done
- Populates the dropdown options once fetching is done
- Selects the first currency in the list once loaded (typically this is
  AUD)
- Sets the conversion rate equal to the selected currency
- Changes the selected currency and the conversion rate string
  dynamically when the user changes the dropdown